Update game1i.txt
I used the ULTIMA282 software application to do the following. Specify a predetermined set of game rules that players agree to at some high stakes blackjack tables found on the Las Vegas strip. Conduct a complete-combinatorial analysis of the card game. Use the results of the complete-combinatorial analysis to assemble a set of combination-dependent-playing-strategy tables for the card game. Use the results of the complete-combinatorial analysis to assemble a combination-dependent-expected-value table for the card game. Use the results of the complete-combinatorial analysis to calculate an estimate of the expected value of the card game. Save the results of the complete-combinatorial analysis to a set of files consisting of a set of data files and a set of text files. 

The predetermined set of game rules specifies the following. The dealer uses a six deck shoe. The dealer stands on soft 17. If the card dealt face up to the dealer's initial hand is not a card selected from a group consisting of an Ace and a card assigned a value of ten points, then the player may surrender the player's initial hand. If the card dealt face up to the dealer's initial hand is a card selected from a group consisting of an Ace and a card assigned a value of ten points,  and if, upon checking the dealer's initial hand for a blackjack,  the dealer does not find a blackjack, then the player may surrender the player's initial hand. The player may split a pair, and re-split a maximum of three times. The player may double down on any two cards. The player may double down after a split. The player may re-split aces. The player may not draw to split aces. Blackjack pays 3 to 2 odds on the game wager.  

The set of text files includes a text file named game3o.txt. This text file contains a set of tables. The set of tables consists of one combination-dependent-playing-strategy table assembled for the set of hands consisting of the player's primary-initial hand and the dealer's initial hand, a composition-dependent-expected-value table assembled for the set of hands consisting of the player's primary-initial hand and the dealer's initial hand, and two combination-dependent-playing-strategy tables assembled for the set of hands consisting of the player's post-split-initial hand and the dealer's initial hand.